requirements.txt
Flask>=1.1,<4.0
numpy>=1.18
scikit-image>=0.16
scikit-learn>=0.23
joblib>=0.16
Pillow>=7.2
pycocotools>=2.0.7
